Bus Time refresh now done by service
Moved all bus times refresh code to a new service
(BusTimeRefreshService)
Updated BusTimeRefreshTask to remove AsyncTask references
Updated android and metawatch code to send/listen for Bus Times intents

+++ src/uk/co/mentalspace/android/bustimes/LocationManager.java
@@ -144,7 +144,7 @@ public class LocationManager {
         }
 	}
 	
-	public static Location getLocationById(Context ctx, int stopId) {
+	public static Location getLocationById(Context ctx, long stopId) {
 		Log.d(LOGNAME, "Getting location by Id: "+stopId);
         LocationsDBAdapter ldba = new LocationsDBAdapter(ctx);
         try {
